Implemented Web Server Class
