+++ src/main.cpp
@@ -33,14 +33,9 @@
  * Includes
  *****************************************************************************/
 
-#include <DNSServer.h>
-#include <ESP8266WebServer.h>
-#include <ESP8266mDNS.h>
-#include <LittleFS.h>
-#include <WebSocketsServer.h>
-
 #include "Board.h"
 #include "WLAN.h"
+#include "WebServer.h"
 
 /******************************************************************************
  * Macros
@@ -63,22 +58,11 @@ typedef enum
 /******************************************************************************
  * Prototypes
  *****************************************************************************/
-
-static void webSocketEvent(uint8_t clientId, WStype_t type, uint8_t *payload, size_t length);
 static void handleCompetition();
 
 /******************************************************************************
  * Variables
  *****************************************************************************/
-
-/** Hostname */
-static const char *HOSTNAME = "laptimer";
-
-/** Webserver port */
-static const uint32_t WEBSERVER_PORT = 80;
-
-/** Websocket port */
-static const uint32_t WEBSOCKET_PORT = 81;
 
 /**
  * After the first detection of the robot with the ext. sensor, this consider
@@ -89,17 +73,14 @@ static const uint32_t SENSOR_BLIND_PERIOD = 400;
 /** Competition start timestamp in ms */
 static uint32_t gStartTimestamp = 0;
 
-/** Webserver on port for http protocol */
-static ESP8266WebServer gWebServer(WEBSERVER_PORT);
-
-/** Websocket server on port for ws protocol */
-static WebSocketsServer gWebSocketSrv(WEBSOCKET_PORT);
-
 /** Current competition state */
 static CompetitionState gCompetitionState = COMPETITION_STATE_UNRELEASED;
 
-/** WLAN instance */
+/** WLAN Instance */
 static WLAN wlan;
+
+/** WebServer Instance */
+static Server webserver;
 
 /******************************************************************************
  * External functions
@@ -117,36 +98,15 @@ void setup()
         Serial.printf("%lu: failed to start Board. \n", millis());
         isError = true;
     }
-    /* Mount filesystem */
-    else if (false == LittleFS.begin())
-    {
-        Serial.printf("%lu: Failed to mount filesystem.\n", millis());
-        isError = true;
-    }
     else if (false == wlan.begin())
     {
         Serial.printf("%lu: Failed to start AP.\n", millis());
         isError = true;
     }
-    /* Setup mDNS service */
-    else if (false == MDNS.begin(HOSTNAME))
+    else if (false == webserver.begin())
     {
-        Serial.printf("%lu: Failed to start mDNS service.", millis());
+        Serial.printf("%lu: Failed to start AP.\n", millis());
         isError = true;
-    }
-    else
-    {
-        /* Setup webserver */
-        gWebServer.serveStatic("/", LittleFS, "/web/", "max-age=86400");
-        gWebServer.onNotFound(
-            []() {
-                gWebServer.send(404, "text/plain", "File not found.");
-            });
-        gWebServer.begin();
-
-        /* Setup websocket server */
-        gWebSocketSrv.onEvent(webSocketEvent);
-        gWebSocketSrv.begin();
     }
 
     if (true == isError)
@@ -166,104 +126,15 @@ void setup()
  */
 void loop()
 {
-    MDNS.update();
-    gWebServer.handleClient();
-    gWebSocketSrv.loop();
+    webserver.cycle();
     handleCompetition();
-
-    return;
 }
 
 /******************************************************************************
  * Local functions
  *****************************************************************************/
 
-/**
- * Handle websocket event.
- *
- * @param[in] clientId  Websocket client id
- * @param[in] type      Event type
- * @param[in] payload   Event payload
- * @param[in] length    Event payload length
- */
-static void webSocketEvent(uint8_t clientId, WStype_t type, uint8_t *payload, size_t length)
-{
-    String cmd;
-    size_t index = 0;
 
-    switch (type)
-    {
-    case WStype_ERROR:
-        Serial.printf("%lu: Ws client (%u) error.\n", millis(), clientId);
-        break;
-
-    case WStype_DISCONNECTED:
-        Serial.printf("%lu: Ws client (%u) disconnected.\n", millis(), clientId);
-        break;
-
-    case WStype_CONNECTED:
-        Serial.printf("%lu: Ws client (%u) connected.\n", millis(), clientId);
-        break;
-
-    case WStype_TEXT:
-        for (index = 0; index < length; ++index)
-        {
-            cmd += reinterpret_cast<char *>(payload)[index];
-        }
-
-        Serial.printf("%lu: Ws client (%u): %s\n", millis(), clientId, cmd.c_str());
-
-        if (cmd.equals("RELEASE"))
-        {
-            if ((COMPETITION_STATE_UNRELEASED == gCompetitionState) ||
-                (COMPETITION_STATE_FINISHED == gCompetitionState))
-            {
-                gCompetitionState = COMPETITION_STATE_RELEASED;
-                gWebSocketSrv.sendTXT(clientId, "ACK");
-            }
-            else
-            {
-                gWebSocketSrv.sendTXT(clientId, "NACK");
-            }
-        }
-        else
-        {
-            gWebSocketSrv.sendTXT(clientId, "NACK");
-        }
-        break;
-
-    case WStype_BIN:
-        /* Not supported. */
-        break;
-
-    case WStype_FRAGMENT_TEXT_START:
-        /* Not supported. */
-        break;
-
-    case WStype_FRAGMENT_BIN_START:
-        /* Not supported. */
-        break;
-
-    case WStype_FRAGMENT:
-        /* Not supported. */
-        break;
-
-    case WStype_FRAGMENT_FIN:
-        /* Not supported. */
-        break;
-
-    case WStype_PING:
-        /* Not supported. */
-        break;
-
-    case WStype_PONG:
-        /* Not supported. */
-        break;
-
-    default:
-        break;
-    }
-}
 
 /**
  * Handle the competition state machine, depending on the user input from
@@ -286,7 +157,7 @@ static void handleCompetition()
         if (true == Board::isRobotDetected())
         {
             gStartTimestamp = millis();
-            gWebSocketSrv.broadcastTXT("EVT;STARTED");
+            webserver.WS_textAll("EVT;STARTED");
             gCompetitionState = COMPETITION_STATE_STARTED;
         }
         break;
@@ -299,7 +170,7 @@ static void handleCompetition()
         {
             if (true == Board::isRobotDetected())
             {
-                gWebSocketSrv.broadcastTXT(String("EVT;FINISHED;") + duration);
+                webserver.WS_textAll(String("EVT;FINISHED;") + duration);
                 gCompetitionState = COMPETITION_STATE_FINISHED;
             }
         }
